+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
+## 4.13.0
+
+### Enhancements
+
+- Adds support for local images and videos in paywalls.
+- Schedules trial notifications after purchasing Stripe products.
+
+### Fixes
+
+- Fixes race condition relating to the user ID when upgrading from v3 of the SDK to v4.
+
 ## 4.12.11
 
 ### Enhancements